requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
alembic
pydantic-settings
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
email-validator
